Allow some pointer wiggle when long-tapping.
That is, only set dragHasMoved if the pointer has moved more than a
trivial amount. (On mobile, you get multiple touch events even if the
user hasn't moved at all, because the pressure varies. But it's also
very hard to avoid moving a pixel or two in the process of making
contact with the device.