Remove ref from effect dependencies
Suggested on review. I think `useRef` should return the same object every time for as long as the DragGroup remains mounted, so there's no upside to including it in the list of dependencies.

I meant to include `dragGroup.current`, but I think that will also return the same DOM element every time for as long as the DragGroup exists.

Co-authored-by: Sarah Edwards <skedwards88@github.com>